feat: 사용자 프로필 조회 기능을 구현한다.
